Move package.json preparation to separate class
